move third party related code into infra folder and seat avalability to appConfig
