+++ pages/api/available-seats.ts
@@ -1,12 +1,12 @@
 import { NextApiHandler } from 'next'
-import { SEAT_AVAILABILITY } from './constants/constants'
 import { db } from './constants/db'
+import { appConfig } from '../../domain/appConfig'
 
 const getAvailableSeats: NextApiHandler = (req, res) => {
   if (req.method === 'GET') {
     return db
       .getSeatAvailability()
-      .then((count) => res.send({ count, maxSeat: SEAT_AVAILABILITY }))
+      .then((count) => res.send({ count, maxSeat: appConfig.seatAvailability }))
       .catch((e) => {
         res.status(500).send(e)
       })
